fix(engine): improved signalling of long polling
Improved implementation of long polling which fixes
a race condition where long polling requests are not
notified if a task is created between the moment in
time where the acquisition thread checks for lockable
tasks and acquires the lock to await a signal.

The improved implementation provides a MPSC condition
implementation based on LockSupport.park()
and LockSupport.unpark(). A call to park() immediateley
returns when "the permit is available" meaning that
unpark() was called since the last time park() returned.
Note that the permit does not accumulate which is what
we want. Also note that the call to park() may spuriously
return which is not a problem in our usecase since
we will just poll for tasks and then continue to wait.

This fixes the race condition: if signal() is called before
the acquisition thread calls await(), the call to await()
will immediately return.

The MPSC condition is wrapped by a Composite condition
which provides a signalAll() method, signalling all
conditions that are currently registered. This effectively
allows multiple acquisition threads to "subscribe" to the
signals. Could be necessary if a user deploys multiple
rest apis on the same engine classpath.

Another advantage of this implementation is that it does
not use locks.

Also adds test verifying that a condition is signalled
when external task is created.

related to #CAM-9196, #298