fix(rest/longpolling): poll periodically
when there are pending requests, the rest API should poll
periodically for situations in which the notifications
cannot work. (Examples: tasks created on other cluster nodes,
tasks with expired timeouts.)

This fix introduces a simple fixed constant at which polling is
being done by the rest API in case it has pending requests.
Default timeout is 30s. More configurability can be added in the
future.

related to CAM-9213, #307