feat: add constructor to aggregate a list of ValidationResults into one
Zeebe uses two different validators, one for design-time validations, and one for run-time validations

Currently, it calls `write` on the two validation results separately. But since we need to limit the size of the printed results, it would be easier to have a way to combine both results in one, and call `write` only once for both.